refactored textinput to a separeate component
