+++ App.js
@@ -2,28 +2,16 @@ import { StatusBar } from 'expo-status-bar';
 import { StyleSheet, Text, TextInput, View } from 'react-native';
 import Header from './components/Header';
 import { useState } from 'react';
+import Input from './components/Input';
 
 export default function App() {
   const appName = "My awesome app";
-  const [text, setText] = useState("");
-
-  //call back handler
-  function changeTextHandler(changedText) {
-    console.log("user is typing", changedText);
-    setText(changedText)
-  }
-
+  
   return (
     <View style={styles.container}>
       <StatusBar style="auto" />
       <Header name={appName} version ={2}/>
-      <TextInput 
-        placeholder="type something"
-        style={styles.input} 
-        value={text} 
-        onChangeText={changeTextHandler}
-      />
-      <Text>{text}</Text>
+      <Input />
     </View>
   );
 }
@@ -35,9 +23,4 @@ const styles = StyleSheet.create({
     alignItems: 'center',
     justifyContent: 'center',
   },
-  input: {
-    borderBottomWidth: 2,
-    borderBottomColor:"purple",
-    width:"50%",
-  }
 });
